Few-shot Learning (#1)
* feat: implement few-shot learning

* feat: cuda

* fix: cuda

* fix: test load

* fix: cuda

* chore: print eval metric

* afd

* fix: pass val step in training

* feat: tqdm in validation

* fix: print sampling count

* feat: 타겟 클래스만 포함하지 않더라도 샘플링

- 타겟 클래스가 아니면 O 태그로 치환

* feat: 샘플링이 안될 경우 샘플링된 support set, query set 초기화

* feat: 샘플링 시 최대 토큰 길이의 태그만 고려

* fix: 메서드 서순

* chore: 샘플 만들때 tqdm 사용

* chore: 타겟 클래스만 포함하도록 샘플링

* fix: 최대 토큰 길이 이후 처리

* fix: 샘플링 스택 에러

* fix: 샘플링 시 target class 처리 에러 수정

* fix: 전체 sample class 갯수 출력

* fix: query max stack

* fix: query max stack

* chore: remove tqdm

+++ trainer.py
@@ -8,7 +8,7 @@ import torch
 from torch.utils.data import DataLoader, RandomSampler, SequentialSampler
 from transformers import AdamW, get_linear_schedule_with_warmup
 
-from utils import compute_metrics, get_labels, get_test_texts, show_report, MODEL_CLASSES
+from utils import compute_metrics, get_labels, get_test_texts, show_report, MODEL_CLASSES, get_labels_from_path
 
 logger = logging.getLogger(__name__)
 
@@ -22,28 +22,30 @@ class Trainer(object):
 
         self.label_lst = get_labels(args)
         self.num_labels = len(self.label_lst)
-        # Use cross entropy ignore index as padding label id so that only real label ids contribute to the loss later
-        self.pad_token_label_id = torch.nn.CrossEntropyLoss().ignore_index
 
-        self.config_class, self.model_class, _ = MODEL_CLASSES[args.model_type]
+        if args.task == 'naver-ner':
+            # Use cross entropy ignore index as padding label id so that only real label ids contribute to the loss later
+            self.pad_token_label_id = torch.nn.CrossEntropyLoss().ignore_index
 
-        self.config = self.config_class.from_pretrained(args.model_name_or_path,
-                                                        num_labels=self.num_labels,
-                                                        finetuning_task=args.task,
-                                                        id2label={str(i): label for i, label in enumerate(self.label_lst)},
-                                                        label2id={label: i for i, label in enumerate(self.label_lst)})
-        self.model = self.model_class.from_pretrained(args.model_name_or_path, config=self.config)
+            self.config_class, self.model_class, _ = MODEL_CLASSES[args.model_type]
 
-        # GPU or CPU
-        self.device = f"cuda:{args.cuda_number}" if torch.cuda.is_available() and not args.no_cuda else "cpu"
-        self.model.to(self.device)
+            self.config = self.config_class.from_pretrained(args.model_name_or_path,
+                                                            num_labels=self.num_labels,
+                                                            finetuning_task=args.task,
+                                                            id2label={str(i): label for i, label in enumerate(self.label_lst)},
+                                                            label2id={label: i for i, label in enumerate(self.label_lst)})
+            self.model = self.model_class.from_pretrained(args.model_name_or_path, config=self.config)
 
-        self.test_texts = None
-        if args.write_pred:
-            self.test_texts = get_test_texts(args)
-            # Empty the original prediction files
-            if os.path.exists(args.pred_dir):
-                shutil.rmtree(args.pred_dir)
+            # GPU or CPU
+            self.device = f"cuda:{args.cuda_number}" if torch.cuda.is_available() and not args.no_cuda else "cpu"
+            self.model.to(self.device)
+
+            self.test_texts = None
+            if args.write_pred:
+                self.test_texts = get_test_texts(args)
+                # Empty the original prediction files
+                if os.path.exists(args.pred_dir):
+                    shutil.rmtree(args.pred_dir)
 
     def train(self):
         train_sampler = RandomSampler(self.train_dataset)
